FB-2 Refactor FizzBuzzApplication
