+++ src/main/java/com/example/fizzbuzz/controller/FizzBuzzController.java
@@ -20,4 +20,6 @@ public class FizzBuzzController {
     public ResponseEntity getAllProducts() {
         return ResponseEntity.ok(this.fizzBuzzRepository.findAll());
     }
+
+
 }
